+++ ResourcesMonitor.Runner/App.xaml
@@ -1,0 +1,9 @@
+﻿<Application x:Class="ResourcesMonitor.Runner.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:local="clr-namespace:ResourcesMonitor.Runner"
+             Startup="Application_Startup">
+    <Application.Resources>
+
+    </Application.Resources>
+</Application>
