+++ ResourcesMonitor.Runner/AssemblyInfo.cs
