+++ ResourcesMonitor.Runner/ResourcesMonitor.Runner.csproj
@@ -1,14 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Library</OutputType>
+    <OutputType>WinExe</OutputType>
     <TargetFramework>net6.0-windows</TargetFramework>
     <Nullable>enable</Nullable>
     <UseWPF>true</UseWPF>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="KitX.Contract.CSharp" Version="22.4.6350.257" />
+    <ProjectReference Include="..\ResourcesMonitor.WPF\ResourcesMonitor.WPF.csproj" />
   </ItemGroup>
 
 </Project>
