+++ ResourcesMonitor.WPF/LoaderStruct.json
