+++ ResourcesMonitor.WPF/MainWindow.xaml
