+++ ResourcesMonitor.WPF/MarketPluginContract.cs
@@ -1,0 +1,8 @@
+﻿using KitX.Contract.CSharp;
+
+namespace ResourceMonitor.WPF;
+
+internal class MarketPluginContract : IMarketPluginContract
+{
+
+}
