+++ ResourcesMonitor.WPF/PluginStruct.json
