+++ ResourcesMonitor.WPF/ResourcesMonitor.WPF.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Library</OutputType>
+    <TargetFramework>net6.0-windows</TargetFramework>
+    <Nullable>enable</Nullable>
+    <UseWPF>true</UseWPF>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="KitX.Contract.CSharp" Version="23.4.6543.429" />
+  </ItemGroup>
+
+</Project>
