+++ ResourcesMonitor.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.3.32825.248
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ResourceMonitor.WPF", "ResourceMonitor.WPF\ResourceMonitor.WPF.csproj", "{EE3C8D87-407D-45CA-B8E7-5BF5B6ABDE03}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ResourcesMonitor.WPF", "ResourcesMonitor.WPF\ResourcesMonitor.WPF.csproj", "{EE3C8D87-407D-45CA-B8E7-5BF5B6ABDE03}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ResourcesMonitor.Runner", "ResourcesMonitor.Runner\ResourcesMonitor.Runner.csproj", "{A1A4EFF7-E2E4-4B07-8AB2-7B70588A2098}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{EE3C8D87-407D-45CA-B8E7-5BF5B6ABDE03}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EE3C8D87-407D-45CA-B8E7-5BF5B6ABDE03}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EE3C8D87-407D-45CA-B8E7-5BF5B6ABDE03}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A1A4EFF7-E2E4-4B07-8AB2-7B70588A2098}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1A4EFF7-E2E4-4B07-8AB2-7B70588A2098}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A1A4EFF7-E2E4-4B07-8AB2-7B70588A2098}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1A4EFF7-E2E4-4B07-8AB2-7B70588A2098}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
